DVFS/PSU: use uint32_t instead of uint64_t for voltage variable.
In this patch voltage variable is change from 64 to 32 bits.
This resolution is more than enough for this implementation.

+++ module/mock_psu/include/mod_mock_psu.h
@@ -92,7 +92,7 @@ struct mod_mock_psu_element_cfg {
     /*!
      * \brief Default voltage, in millivolts (mV), of the device's supply.
      */
-    uint64_t default_voltage;
+    uint32_t default_voltage;
 };
 
 /*!
